requirements.txt
python-dotenv==1.0.0
gradio==3.50.2
langchain==0.0.334
tiktoken==0.5.1
faiss-cpu==1.7.4
torch==2.0.1
sentence-transformers==2.2.2
InstructorEmbedding==1.0.1
openai==1.2.3
loguru==0.7.2
textract==1.6.5
